beta-brede - refactor - removed all uses of old logfing function
