+++ webservice/controller/admin_faq.go
@@ -2,7 +2,6 @@ package controller
 
 import (
 	"github.com/JohanAanesen/CSAMS/webservice/model"
-	"github.com/JohanAanesen/CSAMS/webservice/shared/session"
 	"github.com/JohanAanesen/CSAMS/webservice/shared/view"
 	"log"
 	"net/http"
@@ -11,7 +10,7 @@ import (
 // AdminFaqGET handles GET-request at admin/faq/index
 func AdminFaqGET(w http.ResponseWriter, r *http.Request) {
 	content := model.GetDateAndQuestionsFAQ() // TODO (Svein): Move this to 'settings'
-	if content.Questions == "-1" {            // TODO (Svein): Allow blank FAQ
+	if content.Questions == "-1" { // TODO (Svein): Allow blank FAQ
 		log.Println("Something went wrong with getting the faq (admin.go)")
 		ErrorHandler(w, r, http.StatusInternalServerError)
 		return
@@ -30,7 +29,7 @@ func AdminFaqGET(w http.ResponseWriter, r *http.Request) {
 // AdminFaqEditGET returns the edit view for the faq
 func AdminFaqEditGET(w http.ResponseWriter, r *http.Request) {
 	content := model.GetDateAndQuestionsFAQ() // TODO (Svein): Move this to 'settings'
-	if content.Questions == "-1" {            // TODO (Svein): Allow blank FAQ
+	if content.Questions == "-1" { // TODO (Svein): Allow blank FAQ
 		log.Println("Something went wrong with getting the faq (admin.go)")
 		ErrorHandler(w, r, http.StatusInternalServerError)
 		return
@@ -79,8 +78,10 @@ func AdminFaqUpdatePOST(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	/* TODO brede : log
 	// Get user for logging purposes
 	currentUser := session.GetUserFromSession(r)
+
 
 	// Collect the log data
 	logData := model.Log{
@@ -97,6 +98,8 @@ func AdminFaqUpdatePOST(w http.ResponseWriter, r *http.Request) {
 		ErrorHandler(w, r, http.StatusInternalServerError)
 		return
 	}
+
+	*/
 
 	//AdminFaqGET(w, r)
 	http.Redirect(w, r, "/admin/faq", http.StatusFound)
