+++ webservice/go.mod
@@ -3,6 +3,7 @@ module github.com/JohanAanesen/CSAMS/webservice
 // go 1.12
 
 require (
+	github.com/JohanAanesen/CSAMS/schedulerservice v0.0.0-20190411112816-963fe7b911fc
 	github.com/go-sql-driver/mysql v1.4.1
 	github.com/gorilla/handlers v1.4.0
 	github.com/gorilla/mux v1.7.0
